Resolved broken tests.
