+++ java/test/processing/mode/java/CachedRuntimePathFactoryTest.java
@@ -18,15 +18,12 @@ along with this program; if not, write to the Free Software Foundation, Inc.
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-package processing.mode.java.pdex.util;
+package processing.mode.java;
 
 import org.junit.Before;
 import org.junit.Test;
 import processing.app.Sketch;
-import processing.mode.java.JavaMode;
-import processing.mode.java.pdex.ImportStatement;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
