+++ java/test/processing/mode/java/CodeFolderRuntimePathFactoryTest.java
@@ -18,31 +18,30 @@ along with this program; if not, write to the Free Software Foundation, Inc.
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-package processing.mode.java.pdex.util;
+package processing.mode.java;
 
 import org.junit.Before;
 import org.junit.Test;
 import processing.app.Sketch;
-import processing.mode.java.JavaMode;
-import processing.mode.java.pdex.ImportStatement;
 
+import java.io.File;
 import java.util.List;
 
 import static org.junit.Assert.*;
 
-
-public class ModeSearchRuntimePathFactoryTest {
+public class CodeFolderRuntimePathFactoryTest {
 
   private RuntimePathBuilder.RuntimePathFactoryStrategy factory;
   private JavaMode testMode;
   private List<ImportStatement> testImports;
   private Sketch testSketch;
+
   private List<String> classpath;
 
   @Before
   public void setUp() throws Exception {
     RuntimePathBuilder builder = new RuntimePathBuilder();
-    factory = builder::buildModeSearchPath;
+    factory = builder::buildCodeFolderPath;
     testMode = RuntimePathFactoryTestUtil.createTestJavaMode();
     testImports = RuntimePathFactoryTestUtil.createTestImports();
     testSketch = RuntimePathFactoryTestUtil.createTestSketch();
@@ -51,15 +50,14 @@ public class ModeSearchRuntimePathFactoryTest {
   }
 
   @Test
-  public void buildClasspathLength() {
-    assertEquals(3, classpath.size());
+  public void testBuildClasspathSize() {
+    assertEquals(2, classpath.size());
   }
 
   @Test
-  public void buildClasspathValues() {
-    assertEquals("library6", classpath.get(0));
-    assertEquals("javax.library7", classpath.get(1));
-    assertEquals("library8", classpath.get(2));
+  public void testBuildClasspathValues() {
+    assertEquals("testdir" + File.separator + "file1.jar", classpath.get(0));
+    assertEquals("testdir" + File.separator + "file3.zip", classpath.get(1));
   }
 
 }
