+++ java/test/processing/mode/java/CoreLibraryRuntimePathFactoryTest.java
@@ -18,49 +18,45 @@ along with this program; if not, write to the Free Software Foundation, Inc.
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-package processing.mode.java.pdex.util;
+package processing.mode.java;
 
 import org.junit.Before;
 import org.junit.Test;
 import processing.app.Sketch;
-import processing.mode.java.JavaMode;
-import processing.mode.java.pdex.ImportStatement;
-import processing.mode.java.pdex.util.RuntimePathFactoryTestUtil;
 
-import java.io.File;
 import java.util.List;
 
 import static org.junit.Assert.*;
 
-public class CodeFolderRuntimePathFactoryTest {
+public class CoreLibraryRuntimePathFactoryTest {
 
   private RuntimePathBuilder.RuntimePathFactoryStrategy factory;
   private JavaMode testMode;
   private List<ImportStatement> testImports;
   private Sketch testSketch;
-
-  private List<String> classpath;
+  private List<String> classPath;
 
   @Before
   public void setUp() throws Exception {
     RuntimePathBuilder builder = new RuntimePathBuilder();
-    factory = builder::buildCodeFolderPath;
+    factory = builder::buildCoreLibraryPath;
     testMode = RuntimePathFactoryTestUtil.createTestJavaMode();
     testImports = RuntimePathFactoryTestUtil.createTestImports();
     testSketch = RuntimePathFactoryTestUtil.createTestSketch();
 
-    classpath = factory.buildClasspath(testMode, testImports, testSketch);
+    classPath = factory.buildClasspath(testMode, testImports, testSketch);
   }
 
   @Test
   public void testBuildClasspathSize() {
-    assertEquals(2, classpath.size());
+    assertEquals(2, classPath.size());
   }
 
   @Test
-  public void testBuildClasspathValues() {
-    assertEquals("testdir" + File.separator + "file1.jar", classpath.get(0));
-    assertEquals("testdir" + File.separator + "file3.zip", classpath.get(1));
+  public void testBuildClasspathValue() {
+    assertEquals("library1", classPath.get(0));
+    assertEquals("library2", classPath.get(1));
   }
+
 
 }
