+++ java/test/processing/mode/java/JavaRuntimePathFactoryTest.java
@@ -18,19 +18,18 @@ along with this program; if not, write to the Free Software Foundation, Inc.
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 */
 
-package processing.mode.java.pdex.util;
+package processing.mode.java;
 
 import org.junit.Before;
 import org.junit.Test;
 import processing.app.Sketch;
-import processing.mode.java.JavaMode;
-import processing.mode.java.pdex.ImportStatement;
 
 import java.util.List;
 
 import static org.junit.Assert.*;
 
-public class LibrarySearchRuntimePathFactoryTest {
+
+public class JavaRuntimePathFactoryTest {
 
   private RuntimePathBuilder.RuntimePathFactoryStrategy factory;
   private JavaMode testMode;
@@ -42,7 +41,7 @@ public class LibrarySearchRuntimePathFactoryTest {
   @Before
   public void setUp() throws Exception {
     RuntimePathBuilder builder = new RuntimePathBuilder();
-    factory = builder::buildLibrarySearchPath;
+    factory = builder::buildJavaRuntimePath;
     testMode = RuntimePathFactoryTestUtil.createTestJavaMode();
     testImports = RuntimePathFactoryTestUtil.createTestImports();
     testSketch = RuntimePathFactoryTestUtil.createTestSketch();
@@ -52,14 +51,18 @@ public class LibrarySearchRuntimePathFactoryTest {
 
   @Test
   public void testBuildClasspathSize() {
-    assertEquals(3, classpath.size());
+    assertEquals(RuntimePathBuilder.STANDARD_MODULES.length, classpath.size());
   }
 
   @Test
   public void testBuildClasspathValues() {
-    assertTrue(classpath.get(0).contains("library3"));
-    assertTrue(classpath.get(1).contains("java.library4"));
-    assertTrue(classpath.get(2).contains("library5"));
+    boolean foundTarget = false;
+    for (String entry : classpath) {
+      boolean justFound = entry.contains("java.base.jmod") && entry.contains("jmods");
+      foundTarget = foundTarget || justFound;
+    }
+
+    assertTrue(foundTarget);
   }
 
 }
