+++ java/test/processing/mode/java/SourceUtilTest.java
@@ -1,23 +1,22 @@
-package processing.mode.java.preproc.util;
+package processing.mode.java;
 
 import org.junit.Assert;
 import org.junit.Test;
-import processing.mode.java.preproc.code.SyntaxUtil;
 
 
-public class SyntaxUtilTest {
+public class SourceUtilTest {
 
   @Test
   public void getCountPresent() {
     String input = "test1,test2\n,test3";
-    int count = processing.mode.java.preproc.code.SyntaxUtil.getCount(input, ",");
+    int count = SourceUtil.getCount(input, ",");
     Assert.assertEquals(2, count);
   }
 
   @Test
   public void getCountNotPresent() {
     String input = "test1 test2 test3";
-    int count = SyntaxUtil.getCount(input, ",");
+    int count = SourceUtil.getCount(input, ",");
     Assert.assertEquals(0, count);
   }
